pyproject.toml
[project]
name = "agents-made-easy"
version = "0.1.4"
description = "Reusable lightweight AI agent utilities"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name="Sean Muirhead" }
]
dependencies = [
    "anthropic>=0.75.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
    "claude-agent-sdk>=0.1.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
computer-use = [
    "playwright>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ame"]
